+++ src/main/java/com/workhub/checklist/controller/CheckListController.java
@@ -1,10 +1,10 @@
 package com.workhub.checklist.controller;
 
 import com.workhub.checklist.api.CheckListApi;
-import com.workhub.checklist.dto.CheckListCreateRequest;
-import com.workhub.checklist.dto.CheckListItemStatus;
-import com.workhub.checklist.dto.CheckListResponse;
-import com.workhub.checklist.dto.CheckListUpdateRequest;
+import com.workhub.checklist.dto.checkList.CheckListCreateRequest;
+import com.workhub.checklist.dto.checkList.CheckListItemStatus;
+import com.workhub.checklist.dto.checkList.CheckListResponse;
+import com.workhub.checklist.dto.checkList.CheckListUpdateRequest;
 import com.workhub.checklist.service.checkList.CreateCheckListService;
 import com.workhub.checklist.service.checkList.ReadCheckListService;
 import com.workhub.checklist.service.checkList.UpdateCheckListService;
