+++ src/main/java/com/workhub/checklist/dto/checkList/CheckListItemStatus.java
@@ -1,4 +1,4 @@
-package com.workhub.checklist.dto;
+package com.workhub.checklist.dto.checkList;
 
 public enum CheckListItemStatus {
     PENDING,   // 아직 사용자가 동의/보류 선택 안 함
