+++ src/main/java/com/workhub/checklist/dto/checkList/CheckListOptionFileResponse.java
@@ -1,4 +1,4 @@
-package com.workhub.checklist.dto;
+package com.workhub.checklist.dto.checkList;
 
 import com.workhub.checklist.entity.CheckListOptionFile;
