+++ src/main/java/com/workhub/checklist/dto/checkList/CheckListOptionUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.workhub.checklist.dto;
+package com.workhub.checklist.dto.checkList;
 
 import com.workhub.global.error.ErrorCode;
 import com.workhub.global.error.exception.BusinessException;
@@ -7,21 +7,20 @@ import jakarta.validation.constraints.NotNull;
 
 import java.util.List;
 
-public record CheckListItemUpdateRequest(
-        @NotNull(message = "항목 작업 유형은 필수입니다")
+public record CheckListOptionUpdateRequest(
+        @NotNull(message = "선택지 작업 유형은 필수입니다")
         CheckListUpdateCommandType changeType,
 
-        Long checkListItemId,
-        String itemTitle,
-        Integer itemOrder,
-        Long templateId,
+        Long checkListOptionId,
+        String optionContent,
+        Integer optionOrder,
 
         @Valid
-        List<CheckListOptionUpdateRequest> options
+        List<CheckListOptionFileUpdateRequest> files
 ) {
     /**
      * changeType에 따른 필드 검증
-     * - CREATE: ID 없어야 함, title/order 필수
+     * - CREATE: ID 없어야 함, content/order 필수
      * - UPDATE: ID 필수
      * - DELETE: ID 필수
      */
@@ -32,29 +31,29 @@ public record CheckListItemUpdateRequest(
 
         switch (changeType) {
             case CREATE -> {
-                if (checkListItemId != null) {
+                if (checkListOptionId != null) {
                     throw new BusinessException(ErrorCode.CHECK_LIST_CREATE_CANNOT_HAVE_ID);
                 }
-                if (itemTitle == null || itemOrder == null) {
-                    throw new BusinessException(ErrorCode.CHECK_LIST_CREATE_REQUIRES_TITLE_AND_ORDER);
+                if (optionContent == null || optionOrder == null) {
+                    throw new BusinessException(ErrorCode.CHECK_LIST_CREATE_REQUIRES_CONTENT_AND_ORDER);
                 }
             }
             case UPDATE -> {
-                if (checkListItemId == null) {
+                if (checkListOptionId == null) {
                     throw new BusinessException(ErrorCode.CHECK_LIST_UPDATE_REQUIRES_ID);
                 }
-                // UPDATE는 부분 업데이트 허용: title, order 중 하나만 있어도 OK
+                // UPDATE는 부분 업데이트 허용: content, order 중 하나만 있어도 OK
             }
             case DELETE -> {
-                if (checkListItemId == null) {
+                if (checkListOptionId == null) {
                     throw new BusinessException(ErrorCode.CHECK_LIST_DELETE_REQUIRES_ID);
                 }
             }
         }
 
-        // 하위 options도 검증
-        if (options != null) {
-            options.forEach(CheckListOptionUpdateRequest::validate);
+        // 하위 files도 검증
+        if (files != null) {
+            files.forEach(CheckListOptionFileUpdateRequest::validate);
         }
     }
 }
