+++ src/main/java/com/workhub/checklist/dto/checkList/CheckListUpdateCommandType.java
@@ -1,4 +1,4 @@
-package com.workhub.checklist.dto;
+package com.workhub.checklist.dto.checkList;
 
 /**
  * 체크리스트 업데이트 시 항목/선택지/파일에서 수행할 작업 구분
