+++ src/main/java/com/workhub/checklist/service/checkList/CreateCheckListService.java
@@ -1,12 +1,11 @@
 package com.workhub.checklist.service.checkList;
 
-import com.workhub.checklist.dto.*;
+import com.workhub.checklist.dto.checkList.*;
 import com.workhub.checklist.entity.CheckList;
 import com.workhub.checklist.entity.CheckListItem;
 import com.workhub.checklist.entity.CheckListOption;
 import com.workhub.checklist.entity.CheckListOptionFile;
 import com.workhub.checklist.service.CheckListAccessValidator;
-import com.workhub.checklist.service.CheckListService;
 import com.workhub.global.entity.ActionType;
 import com.workhub.global.error.ErrorCode;
 import com.workhub.global.error.exception.BusinessException;
