+++ src/test/java/com/workhub/checklist/service/chekList/CreateCheckListServiceTest.java
@@ -1,12 +1,12 @@
 package com.workhub.checklist.service.chekList;
 
-import com.workhub.checklist.dto.*;
+import com.workhub.checklist.dto.checkList.*;
 import com.workhub.checklist.entity.CheckList;
 import com.workhub.checklist.entity.CheckListItem;
 import com.workhub.checklist.entity.CheckListOption;
 import com.workhub.checklist.entity.CheckListOptionFile;
 import com.workhub.checklist.service.CheckListAccessValidator;
-import com.workhub.checklist.service.CheckListService;
+import com.workhub.checklist.service.checkList.CheckListService;
 import com.workhub.checklist.service.checkList.CreateCheckListService;
 import com.workhub.global.error.ErrorCode;
 import com.workhub.global.error.exception.BusinessException;
