+++ src/test/java/com/workhub/checklist/service/chekList/UpdateCheckListServiceTest.java
@@ -1,19 +1,19 @@
 package com.workhub.checklist.service.chekList;
 
-import com.workhub.checklist.dto.CheckListDetails;
-import com.workhub.checklist.dto.CheckListItemStatus;
-import com.workhub.checklist.dto.CheckListItemUpdateRequest;
-import com.workhub.checklist.dto.CheckListOptionFileUpdateRequest;
-import com.workhub.checklist.dto.CheckListOptionUpdateRequest;
-import com.workhub.checklist.dto.CheckListResponse;
-import com.workhub.checklist.dto.CheckListUpdateCommandType;
-import com.workhub.checklist.dto.CheckListUpdateRequest;
+import com.workhub.checklist.dto.checkList.CheckListDetails;
+import com.workhub.checklist.dto.checkList.CheckListItemStatus;
+import com.workhub.checklist.dto.checkList.CheckListItemUpdateRequest;
+import com.workhub.checklist.dto.checkList.CheckListOptionFileUpdateRequest;
+import com.workhub.checklist.dto.checkList.CheckListOptionUpdateRequest;
+import com.workhub.checklist.dto.checkList.CheckListResponse;
+import com.workhub.checklist.dto.checkList.CheckListUpdateCommandType;
+import com.workhub.checklist.dto.checkList.CheckListUpdateRequest;
 import com.workhub.checklist.entity.CheckList;
 import com.workhub.checklist.entity.CheckListItem;
 import com.workhub.checklist.entity.CheckListOption;
 import com.workhub.checklist.entity.CheckListOptionFile;
 import com.workhub.checklist.service.CheckListAccessValidator;
-import com.workhub.checklist.service.CheckListService;
+import com.workhub.checklist.service.checkList.CheckListService;
 import com.workhub.checklist.service.checkList.UpdateCheckListService;
 import com.workhub.global.entity.ActionType;
 import com.workhub.global.error.ErrorCode;
